major update
it's been a while and i've made many changes and improvements
i've been exploring an alternate algorithm for traversing the tree which
involves plotting along the ray
also note the vobj old files, old(2) is still in use and contains the
most updated versions of my other rendering method, the rest are just
backups from major changes

+++ msc/functions.h
@@ -10,6 +10,6 @@ void plotlineWithDepth(SDL_Surface* target,double* depth,int x0,int y0,double d0
 
 double sgn(double x);
 
-double pow2(int x);
+double pow2(long long x);
 
 #endif // FUNCTIONS_H_INCLUDED
